Implement serde deserialization for story api
It's important to note that we only want the number of stories, but
because the API returns more than that, and because `serde` refuses to
deserialize any object with extra properties in the json response (bad
serde! shame shame), I've created a struct that includes all the
properties being returned in the response.

Apparently this misbehavior is not followed by `rustc_serialize`, but as
the rustc crate is slated for replacement with serde, I have just bowed
out and allowed serde to serialize the whole thing for no [censored]
reason. There's a bug open against this, but it seems like it's not
getting traction.